LSTM/requirements.txt
tensorflow
numpy==1.26.4
pandas==2.2.2 
scikit-learn==1.4.2
matplotlib==3.8.4
paho-mqtt==2.1.0
pytest==8.2.0
seaborn==0.13.0